move download_directory to base config
