--- mash/services/obs/defaults.py
@@ -1,26 +1,0 @@
-# Copyright (c) 2017 SUSE Linux GmbH.  All rights reserved.
-#
-# This file is part of mash.
-#
-# mash is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# mash is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with mash.  If not, see <http://www.gnu.org/licenses/>
-#
-
-
-class Defaults(object):
-    """
-    Default values
-    """
-    @classmethod
-    def get_download_dir(self):
-        return '/var/lib/mash/images/'
